+++ next-docs/public/examples/icons/Arrows.tsx
@@ -6,6 +6,7 @@ import {
   ArrowsChevronLeftDouble,
   ArrowsChevronRightDouble,
   ArrowsChevronUpDouble,
+  ArrowsCrossLines,
   ArrowsDiagonalsBltr,
   ArrowsDiagonalsTlbr,
   ArrowsDown,
@@ -14,6 +15,7 @@ import {
   ArrowsLeftCurved,
   ArrowsLeftShort,
   ArrowsRefresh,
+  ArrowsRefreshRound,
   ArrowsRemoveBoost,
   ArrowsReply,
   ArrowsRight,
@@ -62,6 +64,9 @@ const Example = ({ wrapperProps, props }: Props) => (
     <Icon {...wrapperProps} name="ArrowsChevronUpDouble">
       <ArrowsChevronUpDouble {...props} />
     </Icon>
+    <Icon {...wrapperProps} name="ArrowsCrossLines">
+      <ArrowsCrossLines {...props} />
+    </Icon>
     <Icon {...wrapperProps} name="ArrowsDiagonalsBltr">
       <ArrowsDiagonalsBltr {...props} />
     </Icon>
@@ -85,6 +90,9 @@ const Example = ({ wrapperProps, props }: Props) => (
     </Icon>
     <Icon {...wrapperProps} name="ArrowsRefresh">
       <ArrowsRefresh {...props} />
+    </Icon>
+    <Icon {...wrapperProps} name="ArrowsRefreshRound">
+      <ArrowsRefreshRound {...props} />
     </Icon>
     <Icon {...wrapperProps} name="ArrowsRemoveBoost">
       <ArrowsRemoveBoost {...props} />