+++ workspaces/icons/src/icons/NotificationsAlert.tsx
@@ -9,8 +9,9 @@ const NotificationsAlert = (props: React.SVGProps<SVGSVGElement>) => (
     {...props}
   >
     <path
-      d="M16.5 20.222a.5.5 0 00-1 0h1zm-1 1.056a.5.5 0 001 0h-1zm1-10.556a.5.5 0 00-1 0h1zm-1 6.334a.5.5 0 001 0h-1zM25 16a9 9 0 01-9 9v1c5.523 0 10-4.477 10-10h-1zm-9 9a9 9 0 01-9-9H6c0 5.523 4.477 10 10 10v-1zm-9-9a9 9 0 019-9V6C10.477 6 6 10.477 6 16h1zm9-9a9 9 0 019 9h1c0-5.523-4.477-10-10-10v1zm-.5 13.222v1.056h1v-1.056h-1zm0-9.5v6.334h1v-6.334h-1z"
-      fill="currentColor"
+      d="M16 20.222v1.056m0-10.556v6.334M25.5 16a9.5 9.5 0 11-19 0 9.5 9.5 0 0119 0z"
+      stroke="currentColor"
+      strokeLinecap="round"
     />
   </svg>
 );