+++ workspaces/icons/src/icons/SportSportsbet.tsx
@@ -14,8 +14,10 @@ const SportSportsbet = (props: React.SVGProps<SVGSVGElement>) => (
       strokeMiterlimit={10}
     />
     <circle
+      cx={3.644}
+      cy={3.644}
       r={3.644}
-      transform="matrix(1 0 0 -1 10.144 10.674)"
+      transform="matrix(1 0 0 -1 6.5 14.318)"
       stroke="currentColor"
       strokeMiterlimit={10}
     />